Initial commit
TopicRouter, Sender(s), {Tcp,Queue}Client

Tests pass, and the thing works with TCP connections. Not all the screws
are perfectly tight yet, so a few more unit tests would be appropriate.

A small problem remains. The 'connection' event of NodeJS's net.server
expects a function that does something with the incoming socket.
Calling socket.address() on it is not returning the peer's address,
however. Instead it returns the socket address of the listening server
(obviously uninteresting). Thus it cannot be used to distinguish clients
in the Senders table. To work around this problem, TcpClient uses a
unique integer as identifier instead (a class variable). Eventually it
would be nice to iron this out and distinguish sockets by their socket
addresses.